requirements.txt
google-generativeai>=0.3.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
pathlib2>=2.3.7
argparse
typing-extensions>=4.8.0
urllib3<2.0.0 